money tagger, hasBefore, hasAfter, looser untag
