+++ .github/workflows/build-and-test.yml
@@ -49,6 +49,6 @@ jobs:
         run: |
           npm run test:smoke
           npm run testb
-          npm run test:perf
+          npm run perf
           npm run test:types
           # npm run test:stress
