+++ .github/workflows/release.yml
@@ -51,7 +51,7 @@ jobs:
         run: |
           npm run test:smoke
           npm run testb
-          npm run test:perf
+          npm run perf
           npm run test:types
           # npm run test:stress
