+++ package.json
@@ -23,9 +23,10 @@
     "testb": "npm run test:smoke && TESTENV=prod  node ./scripts/test/index.js",
     "test:smoke": "node \"./scripts/test/smoke-test/\" | tap-dancer",
     "test:spec": "tape \"./tests/**/*.test.js\" | tap-spec",
-    "test:perf": "node ./scripts/test/speed/index.js",
     "test:types": "ts-node ./scripts/test/types.ts | tap-dancer",
     "test:stress": "node ./scripts/test/stress.js",
+    "perf": "node ./scripts/test/speed/index.js",
+    "perf:build": "node ./scripts/test/speed/build-speed.js",
     "coverage": "node ./scripts/test/coverage.js",
     "coverage:html": "nyc --reporter=html tape \"./tests/**/*.test.js\" | tap-dancer --color always",
     "lint": "eslint ./src/ && eslint ./plugins/**/src/",
