+++ plugins/numbers/src/tagger.js
@@ -12,5 +12,12 @@ const tagger = function(doc) {
   doc.match('#Value and a (half|quarter)').tag('Value', 'value-and-a-half')
   //one hundred and seven dollars
   doc.match('#Money and #Money #Currency?').tag('Money', 'money-and-money')
+
+  // $5.032 is invalid money
+  doc
+    .match('#Money')
+    .not('#TextValue')
+    .match('/\\.[0-9]{3}$/')
+    .unTag('#Money', 'three-decimal money')
 }
 module.exports = tagger
