+++ src/02-tagger/04-correction/matches/06-value.js
@@ -6,24 +6,24 @@ module.exports = [
   { match: '#NumericValue #PhoneNumber', tag: 'PhoneNumber', reason: '(800) PhoneNumber' },
 
   // ==== Currency ====
+  // chinese yuan
   { match: '#Demonym #Currency', tag: 'Currency', reason: 'demonym-currency' },
   // ==== Ordinal ====
   { match: '[second] #Noun', group: 0, tag: 'Ordinal', reason: 'second-noun' },
-  // ==== Money ====
-  { match: '#Value+ #Currency', tag: 'Money', reason: '15 usd' },
   // ==== Unit ====
   //5 yan
   { match: '#Value+ [#Currency]', group: 0, tag: 'Unit', reason: '5-yan' },
   { match: '#Value [(foot|feet)]', group: 0, tag: 'Unit', reason: 'foot-unit' },
-
+  // ==== Money ====
+  { match: '[#Value+] #Currency', group: 0, tag: 'Money', reason: '15 usd' },
   //minus 7
   { match: '(minus|negative) #Value', tag: 'Value', reason: 'minus-value' },
   //5 kg.
   { match: '#Value #Abbreviation', tag: 'Value', reason: 'value-abbr' },
   //seven point five
   { match: '#Value (point|decimal) #Value', tag: 'Value', reason: 'value-point-value' },
-  // ten grand
-  { match: '#Value grand', tag: 'Value', reason: 'value-grand' },
+  // ten bucks
+  { match: '(#Value|a) [(buck|bucks|grand)]', group: 0, tag: 'Currency', reason: 'value-bucks' },
   //quarter million
   { match: '#Determiner [(half|quarter)] #Ordinal', group: 0, tag: 'Value', reason: 'half-ordinal' },
   { match: 'a #Value', tag: 'Value', reason: 'a-value' },
